move GCS (and future AWS) dependencies into separate paths
- Not all EDPs/MPs will want to run on any specific cloud
- This means that unused libraries can be ignored, rather than required
for generic installations
- DefaultStorageFactories renamed to ExampleStorageFactories as it is
only used by example code
- Example code moved into a deployment subdirectory, again, to more
easily isolate optional dependencies.
