+++ src/main/kotlin/org/wfanet/panelmatch/client/deploy/example/gcloud/BUILD.bazel
@@ -5,26 +5,28 @@ load("@rules_java//java:defs.bzl", "java_binary")
 package(default_visibility = ["//visibility:private"])
 
 kt_jvm_library(
-    name = "filesystem",
+    name = "gcloud",
     srcs = glob(["*.kt"]),
     deps = [
         "//src/main/kotlin/org/wfanet/panelmatch/client/deploy",
+        "//src/main/kotlin/org/wfanet/panelmatch/client/deploy/example",
         "//src/main/kotlin/org/wfanet/panelmatch/client/storage",
         "//src/main/kotlin/org/wfanet/panelmatch/common/secrets",
         "@wfa_common_jvm//imports/java/picocli",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/gcs",
     ],
 )
 
 java_binary(
-    name = "FilesystemExampleDaemonMain",
-    main_class = "org.wfanet.panelmatch.client.deploy.filesystem.FilesystemExampleDaemonMainKt",
-    runtime_deps = [":filesystem"],
+    name = "GoogleCloudExampleDaemonMain",
+    main_class = "org.wfanet.panelmatch.client.deploy.gcloud.GoogleCloudExampleDaemonMainKt",
+    runtime_deps = [":gcloud"],
 )
 
 java_image(
-    name = "filesystem_example_daemon_image",
+    name = "google_cloud_example_daemon_image",
     data = ["@grpc_health_probe//file"],
-    main_class = "org.wfanet.panelmatch.client.deploy.filesystem.FilesystemExampleDaemonMainKt",
-    runtime_deps = [":filesystem"],
+    main_class = "org.wfanet.panelmatch.client.deploy.gcloud.GoogleCloudExampleDaemonMainKt",
+    runtime_deps = [":gcloud"],
 )
