+++ src/main/kotlin/org/wfanet/panelmatch/client/storage/aws/s3/S3StorageFactory.kt
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.storage
+package org.wfanet.panelmatch.client.storage.aws.s3
 
 import org.wfanet.measurement.storage.StorageClient
+import org.wfanet.panelmatch.client.storage.StorageDetails
+import org.wfanet.panelmatch.client.storage.StorageFactory
 import org.wfanet.panelmatch.common.ExchangeDateKey
 
 class S3StorageFactory(
