+++ src/main/kotlin/org/wfanet/panelmatch/client/storage/gcloud/gcs/BUILD.bazel
@@ -3,24 +3,19 @@ load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
 package(default_visibility = ["//visibility:public"])
 
 kt_jvm_library(
-    name = "storage",
+    name = "gcs",
     srcs = glob(["*.kt"]),
     deps = [
+        "//imports/java/com/google/cloud/core",
+        "//imports/java/com/google/common:guava",
         "//src/main/kotlin/org/wfanet/panelmatch/client/common",
-        "//src/main/kotlin/org/wfanet/panelmatch/common",
-        "//src/main/kotlin/org/wfanet/panelmatch/common/certificates",
-        "//src/main/proto/wfa/measurement/api/v2alpha:data_provider_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha:exchange_step_attempt_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha:exchange_step_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha:exchange_workflow_kt_jvm_proto",
+        "//src/main/kotlin/org/wfanet/panelmatch/client/storage",
         "//src/main/proto/wfa/panelmatch/client/exchangetasks:exchangesteps_kt_jvm_proto",
         "//src/main/proto/wfa/panelmatch/client/storage:storage_details_kt_jvm_proto",
+        "@wfa_common_jvm//imports/java/com/google/cloud/storage",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
         "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:hashing",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:signatures",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:signed_blob",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/gcs",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage:client",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/filesystem:client",
